Fix more compiler warnings.

+++ Source/SPProcessListController.m
@@ -468,7 +468,7 @@ static const NSString *SPTableViewIDColumnIdentifier = @"Id";
  */
 - (id)tableView:(NSTableView *)tableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(NSInteger)row
 {	
-	id object = (row < [processesFiltered count]) ? [[processesFiltered objectAtIndex:row] valueForKey:[tableColumn identifier]] : @"";
+	id object = ((NSUInteger)row < [processesFiltered count]) ? [[processesFiltered objectAtIndex:row] valueForKey:[tableColumn identifier]] : @"";
 		
 	return (![object isNSNull]) ? object : [prefs stringForKey:SPNullValue];
 }
@@ -624,7 +624,7 @@ static const NSString *SPTableViewIDColumnIdentifier = @"Id";
 - (void)_updateSelectedAutoRefreshIntervalInterface
 {	
 	BOOL found = NO;
-	NSUInteger interval = [prefs integerForKey:SPProcessListAutoRrefreshInterval];
+	NSInteger interval = [prefs integerForKey:SPProcessListAutoRrefreshInterval];
 	
 	NSArray *items = [[autoRefreshIntervalMenuItem submenu] itemArray];
